refactor(subagents): split task tool orchestration
